pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "code-route"
version = "0.1.0"
description = "A self-improving AI assistant framework with dynamic tool creation and orchestration"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Mihajlo Micic", email = "mihajlo@example.com" }]
keywords = ["ai", "assistant", "tools", "cli", "automation", "self-improving"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "anthropic", # Keeping for backward compatibility
    "openai>=1.0.0", # For OpenRouter integration
    "python-dotenv",
    "rich",
    "requests",
    "beautifulsoup4",
    "validators>=0.34.0",
    "PyAutoGUI",
    "Pillow",
    "prompt-toolkit",
    "matplotlib>=3.9.2",
    "flask>=3.0.3",
    "werkzeug>=3.1.2",
    "markdownify>=0.14.1",
    "protego>=0.3.1",
    "readability-lxml>=0.8.1",
    "e2b-code-interpreter>=1.0.3",
    "streamlit>=1.45.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "ruff",
    "mypy",
    "pre-commit",
    "types-requests",
]

[project.urls]
Homepage = "https://github.com/micic-mihajlo/code-route"
Repository = "https://github.com/micic-mihajlo/code-route"
Documentation = "https://github.com/micic-mihajlo/code-route#readme"
"Bug Tracker" = "https://github.com/micic-mihajlo/code-route/issues"

[tool.ruff]
line-length = 88
target-version = "py39"
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "RUF", # ruff-specific rules
]
ignore = []

[tool.ruff.isort]
known-first-party = ["cr"]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.black]
line-length = 88
target-version = ["py39"]
include = '\.pyi?$'

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = ["wikipedia.*", "pyautogui.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=cr --cov-report=term-missing"
testpaths = ["tests"]

[tool.uv.workspace]
members = ["testfolder"]

[project.scripts]
code-route = "code_route.cli:main"
cr = "code_route.cli:main"
